(refactor) idesc_table: Merge headers again
Merge src/kernel/task/resource/idesc_table.h into
      src/include/kernel/task/resource/idesc_table.h

They are being used together anyway, so it's better to merge them
